requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
mysql-connector-python
